refactor the swap cost calculator, more precise access modifiers to variables and functions
